+++ project-editor/src/renderer/components/Modals/SettingsTabs/AppSettings/AppPlaygroundSettings.vue
@@ -43,7 +43,7 @@ export default {
     },
 
     components: { CustomInput: CustomInputs },
-    computed: mapGetters('AppData', ['playgroundSizes', 'autoPlaygroundSizesEnabled']),
+    computed: mapGetters('EditorData', ['playgroundSizes', 'autoPlaygroundSizesEnabled']),
     mounted() {
         this.playgroundWidth = this.playgroundSizes[0];
         this.playgroundHeight = this.playgroundSizes[1];